requirements.txt
requests
python-dotenv
openai
tiktoken
beautifulsoup4
pandas
openpyxl